build: introduce rust
Upcoming patches will simplify xdiff, while also porting parts of it to
Rust. In preparation, add some stubs and setup the Rust build. For now,
it is easier to let cargo build rust and have make or meson merely link
against the static library that cargo builds. In line with ongoing
libification efforts, use multiple crates to allow more modularity on
the Rust side. The xdiff and ivec crates are the focal point for this
series.

cbindgen will be used to generate header files for each Rust crate for
C interoperability.

Signed-off-by: Ezekiel Newren <ezekielnewren@gmail.com>